Update Lander
Lander is created, it works to draw.
I still need to apply the rules to make it work.
Such as getting the thrusters, gravity and Inertia to work. But for now, it will work to play the game at the very least.

I would also like to create an above and beyond part where I could make the program catch rings which would give the end-user more points.